requirements.txt
# Cheminformatics Core
rdkit-pypi>=2022.3.1

# Data Manipulation
pandas>=1.3.0
numpy>=1.21.0

# Visualization & Notebook support
ipython>=7.20.0
matplotlib>=3.4.0
